feat: Update tasks, client code, and added rdflib dependency
Enhanced tasks.py with more verbose code comments, better error handling and specific typing hints. Added RDFLib as to generate RDF Hextuples output instead of a dictionary. renamed `meta` to `rdf` and added the mimetype to support multiple formats to parse RDF output. Added D1985 for investigating `1985` kind events. This required a rework of client.py to include the introduction of several new features for selecting different kinds to subscribe to, filtering of events in the handlers using a `StopProcessingHandlers` exception, better evaluation when to decrypt or not, and better function parameter definitions for broadcast and listen_forever. Added 'rdflib' to the project dependencies for handling RDF graphs.